requirements.txt
amqp==5.1.1
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==37.0.4
daphne==3.0.2
dateutils==0.6.12
dj-database-url==1.0.0
Django==4.0.7
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-phonenumber-field==6.3.0
django-reset-migrations==0.4.0
django-rest-framework-braces==0.3.4
django-scheduler==0.9.6
django-timezone-field==5.0
django-twilio==0.14.0
djangorestframework==3.13.1
hyperlink==21.0.0
icalendar==4.1.0
idna==3.3
incremental==21.3.0
kombu==5.2.4
phonenumbers==8.12.54
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyOpenSSL==22.0.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
twilio==7.12.1
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0
